Fixed the boundary generator.

+++ Engine/Source/Core/Structures/BoundingSphere.cpp
@@ -2,25 +2,25 @@
 #include "BoundingSphere.h"
 
 namespace fe {
-	MeshBoundingSphereHierarchy::MeshBoundingSphereHierarchy(const std::vector<glm::vec3>& vertices, const std::vector<glm::ivec3>& faces)
+	MeshBoundingSphereHierarchy::MeshBoundingSphereHierarchy(const std::vector<glm::dvec3>& vertices, const std::vector<glm::ivec3>& faces)
 		: Tree<BoundingSphere>(faces.size()), m_Vertices(vertices), m_Faces(faces), m_TriangleCenters(faces.size())
 	{
 		std::ranges::transform(m_Faces.begin(), m_Faces.end(), m_TriangleCenters.begin(),
 		[&](const glm::ivec3& f)
 		{
-			return 1.0f / 3.0f * (m_Vertices[f[0]] + m_Vertices[f[1]] + m_Vertices[f[2]]);
+			return 1.0 / 3.0 * (m_Vertices[f[0]] + m_Vertices[f[1]] + m_Vertices[f[2]]);
 		});
 	}
 
-	const glm::vec3& MeshBoundingSphereHierarchy::GetEntityPosition(uint32_t index) const
+	const glm::dvec3& MeshBoundingSphereHierarchy::GetEntityPosition(unsigned int index) const
 	{
 		return m_TriangleCenters[index];
 	}
 
-	void MeshBoundingSphereHierarchy::Calculate(const uint32_t b, const uint32_t n, BoundingSphere& hull) const
+	void MeshBoundingSphereHierarchy::Calculate(const unsigned int b, const unsigned int n, BoundingSphere& hull) const
 	{
-		std::vector<glm::vec3> vertexSubset = std::vector<glm::vec3>(3 * n);
-		for (uint32_t i(0); i < n; ++i)
+		std::vector<glm::dvec3> vertexSubset = std::vector<glm::dvec3>(3 * n);
+		for (unsigned int i(0); i < n; ++i)
 		{
 			const auto& f = m_Faces[m_List[b + i]];
 			{
